refactor(yunxin): move app.systemUIHint to app.ts, narrow type to "question"
